KALMAN: Adjusted covariances.
Eventually achieved best results by increasing the filter’s sensitivity
(adding process/equation noise), but reducing verification distance.
The alternative was to reduce process noise - essentially saying “the
equations are right, listen to them and ignore the data” which caused
the filter to stubbornly ignore the trajectory detections, but required
v_dist to be ranked up.